Sparse Array Excercise
A sparse array is an array of data in which many elements have a value of zero. This is in contrast to a dense array, where most of the elements have non-zero values or are “full” of numbers. A sparse array may be treated differently than a dense array in digital data handling.
4
aba
baba
aba
xzxb
3
aba
xzxb
ab